# api/assets/swagger.json
{
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "GK Circle PCS examination API",
    "version": "0.0.1-alpha"
  },
  "host": "127.0.0.1",
  "basePath": "/api",
  "paths": {
...
    "/v1/user_played_quizes": {
      "get": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["UserPlayedQuiz"],
        "summary": "List played quiz by user.",
        "operationId": "UserPlayedQuiz",
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseListUserPlayedQuizes"
          },
          "401": {
            "$ref": "#/responses/GenericResFailConflict"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
    "/v1/user_played_quizes/{invitationCode}": {
      "post": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["UserPlayedQuiz"],
        "summary": "Validate the code and copy the questions in reponses for user.",
        "operationId": "RequestPlayedQuizValidation",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "InvitationCode",
            "name": "invitationCode",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponsePlayedQuizValidation"
          },
          "401": {
            "$ref": "#/responses/GenericResFailConflict"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
    "/v1/user_played_quizes/{user_played_quiz_id}": {
      "get": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["UserPlayedQuiz"],
        "summary": "Analysis of played quiz by user.",
        "operationId": "RequestListUserPlayedQuizesWithQuestionById",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "UserPlayedQuizId",
            "name": "user_played_quiz_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseListUserPlayedQuizesWithQuestionById"
          },
          "401": {
            "$ref": "#/responses/GenericResFailConflict"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    }
  },
  "definitions": {
    "NullString": {
      "description": "var s NullString\nerr := db.QueryRow(\"SELECT name FROM foo WHERE id=?\", id).Scan(\u0026s)\n...\nif s.Valid {\nuse s.String\n} else {\nNULL value\n}",
      "type": "object",
      "title": "NullString represents a string that may be null.\nNullString implements the Scanner interface so\nit can be used as a scan destination:",
      "properties": {
        "String": {
          "type": "string"
        },
        "Valid": {
          "type": "boolean"
        }
      },
      "x-go-package": "database/sql"
    },
...
  },
  "responses": {
    "GenericResError": {
      "description": "Unexpected error occurred",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "status": {
            "type": "string",
            "enum": ["error"],
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "GenericResFailConflict": {
      "description": "Server understand request but refuse to authorize it",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "enum": ["fail"],
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseListUserPlayedQuizesWithQuestionById": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "calculated_points": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "CalculatedPoints"
                },
                "calculated_score": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "CalculatedScore"
                },
                "correct_answer": {
                  "type": "string",
                  "x-go-name": "CorrectAnswer"
                },
                "is_attend": {
                  "type": "boolean",
                  "x-go-name": "IsAttend"
                },
                "options": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "x-go-name": "Options"
                },
                "options_media": {
                  "type": "string",
                  "x-go-name": "OptionsMedia"
                },
                "points": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "Points"
                },
                "question": {
                  "type": "string",
                  "x-go-name": "Question"
                },
                "question_media": {
                  "type": "string",
                  "x-go-name": "QuestionsMedia"
                },
                "question_type": {
                  "type": "string",
                  "x-go-name": "QuestionType"
                },
                "question_type_id": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "QuestionTypeID"
                },
                "raw_options": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "uint8"
                  },
                  "x-go-name": "RawOptions"
                },
                "resource": {
                  "type": "string",
                  "x-go-name": "Resource"
                },
                "response_time": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "ResponseTime"
                },
                "selected_answer": {
                  "$ref": "#/definitions/NullString"
                }
              }
            },
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponsePlayedQuizValidation": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "session_id": {
                  "type": "string",
                  "x-go-name": "SessionId"
                },
                "user_played_quiz": {
                  "type": "string",
                  "x-go-name": "UserPlayedQuizId"
                }
              }
            },
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
  }
}